+++ pennylane/templates/layers/random.py
@@ -25,7 +25,6 @@ from pennylane.templates.utils import (
     get_shape,
 )
 from pennylane.wires import Wires
-from pennylane.numpy import tensor
 
 
 def random_layer(weights, wires, ratio_imprim, imprimitive, rotations, seed):
@@ -51,10 +50,7 @@ def random_layer(weights, wires, ratio_imprim, imprimitive, rotations, seed):
             gate = np.random.choice(rotations)
             rnd_wire = wires.select_random(1)
 
-            if isinstance(weights[i], tensor):
-                gate(weights[i].unwrap(), wires=rnd_wire)
-            else:
-                gate(weights[i], wires=rnd_wire)
+            gate(weights[i], wires=rnd_wire)
 
             i += 1
         else: