+++ tests/test_optimize_qng.py
@@ -40,7 +40,7 @@ class TestExceptions:
         params = 0.5
 
         with pytest.raises(
-            ValueError, match="The objective function must either be encoded as a single QNode or a VQECost object"
+            ValueError, match="The objective function must either be encoded as a single QNode or an ExpvalCost object"
         ):
             opt.step(cost, params)
 
@@ -143,7 +143,7 @@ class TestOptimize:
         assert np.allclose(cost_fn(theta), -1.41421356, atol=tol, rtol=0)
 
     def test_single_qubit_vqe_using_vqecost(self, tol):
-        """Test single-qubit VQE using VQECost 
+        """Test single-qubit VQE using ExpvalCost
         has the correct QNG value every step, the correct parameter updates,
         and correct cost after 200 steps"""
         dev = qml.device("default.qubit", wires=1)
@@ -160,7 +160,7 @@ class TestOptimize:
 
         h = qml.Hamiltonian(coeffs=coeffs, observables=obs_list)
 
-        cost_fn = qml.VQECost(ansatz=circuit, hamiltonian=h, device=dev)
+        cost_fn = qml.ExpvalCost(ansatz=circuit, hamiltonian=h, device=dev)
 
         def gradient(params):
             """Returns the gradient"""